继续添加内容
1. 更新所有价格为flat64
2. 新增获取交易深度信息
3. 新增获取交易细节信息

+++ models/KLine.go
@@ -4,10 +4,10 @@ type KLineData struct {
 	ID     int64   `json:"id"`     // K线ID
 	Amount float64 `json:"amount"` // 成交量
 	Count  int64   `json:"count"`  // 成交笔数
-	Open   float32 `json:"open"`   // 开盘价
-	Close  float32 `json:"close"`  // 收盘价, 当K线为最晚的一根时, 时最新成交价
-	Low    float32 `json:"low"`    // 最低价
-	High   float32 `json:"high"`   // 最高价
+	Open   float64 `json:"open"`   // 开盘价
+	Close  float64 `json:"close"`  // 收盘价, 当K线为最晚的一根时, 时最新成交价
+	Low    float64 `json:"low"`    // 最低价
+	High   float64 `json:"high"`   // 最高价
 	Vol    float64 `json:"vol"`    // 成交额, 即SUM(每一笔成交价 * 该笔的成交数量)
 }
